requirements.txt
Django==1.8.4
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.5.1
psycopg2==2.7.4
psycopg2-binary==2.7.4

